Update field def

+++ habitica_slack/models.py
@@ -2,4 +2,4 @@ from django.db import models
 
 
 class LastPostTimeStamp(models.Model):
-    time_stamp = models.IntegerField()
+    time_stamp = models.BigIntegerField()
